+++ Engine/Types/ClearType.cs
@@ -1,4 +1,4 @@
 ﻿namespace Exodrifter.Rumor.Engine
 {
-	public enum ClearType { All = 3, Choices = 1, Dialog = 2 }
+	public enum ClearType { All = 3, Choices = 1, Dialog = 2, Choice = 4 }
 }